+++ IcingTaskManager@json/files/IcingTaskManager@json/3.2/specialMenus.js
@@ -489,7 +489,7 @@ AppThumbnailHoverMenu.prototype = {
     this.app = parent.app;
     this.isFavoriteApp = parent.isFavoriteApp;
     this.appList = parent.appList;
-    this._tooltip = new Tooltips.PanelItemTooltip(this._applet, '', parent.orientation);
+    this._tooltip = new Tooltips.PanelItemTooltip(parent, '', this._applet.orientation);
 
     this.actor.hide();
     Main.layoutManager.addChrome(this.actor, {});
@@ -505,7 +505,7 @@ AppThumbnailHoverMenu.prototype = {
     this.signals.connect(this.actor, 'key-release-event', Lang.bind(this, this._onKeyRelease));
   },
 
-  _onButtonPress: function (actor) {
+  _onButtonPress: function () {
     if (this._applet.onClickThumbs && this.appSwitcherItem.box.get_children().length > 1) {
       return;
     }
@@ -558,7 +558,6 @@ AppThumbnailHoverMenu.prototype = {
     if (this.metaWindows.length === 0) {
       this._tooltip.set_text(this.appGroup.appName);
       this._tooltip.show();
-      this._tooltip.preventShow = false;
     } else {
       setTimeout(()=>this.appSwitcherItem._refresh(), 0);
       PopupMenu.PopupMenu.prototype.open.call(this, this._applet.animateThumbs);
@@ -569,7 +568,6 @@ AppThumbnailHoverMenu.prototype = {
     if (this.metaWindows.length === 0) {
       this._tooltip.set_text('');
       this._tooltip.hide();
-      this._tooltip.preventShow = true;
     } else {
       PopupMenu.PopupMenu.prototype.close.call(this, this._applet.animateThumbs);
     }
@@ -625,46 +623,50 @@ PopupMenuAppSwitcherItem.prototype = {
 
   _onKeyPress: function(actor, e){
     let symbol = e.get_key_symbol();
-    let i = _.findIndex(this.appThumbnails, (thumb)=>{
-      return thumb.thumbnail.entered;
+    let i = _.findIndex(this.appThumbnails, (thumbnail)=>{
+      return thumbnail.entered;
     });
 
     let entered = i > -1;
     if (entered) {
-      this.appThumbnails[i].thumbnail.handleLeaveEvent();
+      this.appThumbnails[i].handleLeaveEvent();
     } else {
-      i = _.findIndex(this.appThumbnails, (thumb)=>{
-        return thumb.thumbnail._hasFocus();
+      i = _.findIndex(this.appThumbnails, (thumbnail)=>{
+        return thumbnail._hasFocus();
       });
       if (i === -1) {
         i = 0;
       }
     }
-    let args = this._applet.verticalThumbs ? [Clutter.KEY_Up, Clutter.KEY_Down] : [Clutter.KEY_Left, Clutter.KEY_Right];
+    let args;
     let closeArg;
     if (this._applet.orientation === St.Side.TOP) {
       closeArg = Clutter.KEY_Up;
+      args = [Clutter.KEY_Left, Clutter.KEY_Right];
     } else if (this._applet.orientation === St.Side.BOTTOM) {
       closeArg = Clutter.KEY_Down;
+      args = [Clutter.KEY_Right, Clutter.KEY_Left];
     } else if (this._applet.orientation === St.Side.LEFT) {
       closeArg = Clutter.KEY_Left;
+      args = [Clutter.KEY_Up, Clutter.KEY_Down];
     } else if (this._applet.orientation === St.Side.RIGHT) {
       closeArg = Clutter.KEY_Right;
+      args = [Clutter.KEY_Down, Clutter.KEY_Up];
     }
     let index;
     if (symbol === args[0]) {
       if (!entered) {
         index = i;
-      } else if (this.appThumbnails[i+1] !== undefined) {
-        index = i+1;
+      } else if (this.appThumbnails[i + 1] !== undefined) {
+        index = i + 1;
       } else {
         index = 0;
       }
     } else if (symbol === args[1]) {
       if (!entered) {
         index = i;
-      } else if (this.appThumbnails[i-1] !== undefined) {
-        index = i-1;
+      } else if (this.appThumbnails[i - 1] !== undefined) {
+        index = i - 1;
       } else {
         index = this.appThumbnails.length - 1;
       }
@@ -677,7 +679,7 @@ PopupMenuAppSwitcherItem.prototype = {
       return;
     }
     if (this.appThumbnails[index] !== undefined) {
-      this.appThumbnails[index].thumbnail.handleEnterEvent();
+      this.appThumbnails[index].handleEnterEvent();
     }
   },
 
@@ -696,7 +698,7 @@ PopupMenuAppSwitcherItem.prototype = {
       this.metaWindowThumbnail.setMetaWindow(metaWindow, metaWindows);
     }
     for (let w = 0, len = this.appThumbnails.length; w < len; w++) {
-      this.appThumbnails[w].thumbnail.setMetaWindow(null, metaWindows);
+      this.appThumbnails[w].setMetaWindow(null, metaWindows);
     }
   },
 
@@ -731,7 +733,7 @@ PopupMenuAppSwitcherItem.prototype = {
     let windows = this.metaWindows;
 
 
-    if (this.metaWindowThumbnail && this.metaWindowThumbnail.thumbnail != null) {
+    if (this.metaWindowThumbnail && this.metaWindowThumbnail != null) {
       this.metaWindowThumbnail = null;
     }
     if (this.metaWindowThumbnail && _.isEqual(this.metaWindowThumbnail.metaWindow, this.metaWindow)) {
@@ -749,27 +751,27 @@ PopupMenuAppSwitcherItem.prototype = {
     setTimeout(()=>this.setStyleOptions(windows), 0);
     if (refreshThumbnails) {
       for (let i = 0, len = this.appThumbnails.length; i < len; i++) {
-        this.appThumbnails[i].thumbnail._refresh(windows[0], windows);
+        this.appThumbnails[i]._refresh(windows[0], windows);
       }
     }
   },
 
   _refreshThumbnails: function(){
     for (let i = 0, len = this.appThumbnails.length; i < len; i++) {
-      if (this.appThumbnails[i] !== undefined && this.appThumbnails[i].thumbnail) {
-        this.appThumbnails[i].thumbnail._refresh();
+      if (this.appThumbnails[i] !== undefined && this.appThumbnails[i]) {
+        this.appThumbnails[i]._refresh();
       }
     }
   },
 
   addWindowThumbnails: function (metaWindows) {
+    if (!metaWindows) {
+      metaWindows = this.metaWindows;
+    }
     if (metaWindows.length > 0) {
       let children = this.box.get_children();
       for (let w = 0, len = children.length; w < len; w++) {
         this.box.remove_actor(children[w]);
-      }
-      if (this._applet.sortThumbs) {
-        this.appThumbnails = _.orderBy(this.appThumbnails, ['metaWindow.user_time'], ['asc']);
       }
       this.reAdd = true;
     }
@@ -777,11 +779,7 @@ PopupMenuAppSwitcherItem.prototype = {
     for (let i = 0, len = metaWindows.length; i < len; i++) {
       let metaWindow = metaWindows[i];
       let refThumb = _.findIndex(this.appThumbnails, {metaWindow: metaWindow});
-      if (this.appThumbnails[i] !== undefined && this.appThumbnails[i] && refThumb !== -1) {
-        if (this.reAdd) {
-          this.box.insert_actor(this.appThumbnails[i].thumbnail.actor, 0);
-        }
-      } else {
+      if (!this.appThumbnails[i] && refThumb === -1) {
         if (this.metaWindowThumbnail) {
           this.metaWindowThumbnail.destroy(true);
         }
@@ -791,13 +789,17 @@ PopupMenuAppSwitcherItem.prototype = {
           metaWindows: metaWindows
         });
         thumbnail.setMetaWindow(metaWindow, metaWindows);
-        this.appThumbnails.push({
-          metaWindow: metaWindow,
-          thumbnail: thumbnail
-        });
-        this.box.insert_actor(this.appThumbnails[i].thumbnail.actor, 0);
+        this.appThumbnails.push(thumbnail);
       }
     }
+    if (this._applet.sortThumbs) {
+      this.appThumbnails = _.orderBy(this.appThumbnails, (thumbnail)=>{
+        return thumbnail.metaWindow.user_time;
+      }, ['desc']);
+    }
+    each(this.appThumbnails, (thumbnail, i)=>{
+      this.box.insert_actor(thumbnail.actor, i);
+    });
     this.box.show();
   },
   setStyleOptions: function (windows) {
@@ -818,8 +820,8 @@ PopupMenuAppSwitcherItem.prototype = {
       return;
     }
     for (let i = 0, len = this.appThumbnails.length; i < len; i++) {
-      if (this.appThumbnails[i].thumbnail) {
-        this.appThumbnails[i].thumbnail.thumbnailIconSize();
+      if (this.appThumbnails[i]) {
+        this.appThumbnails[i].thumbnailIconSize();
       }
     }
   },
@@ -827,9 +829,9 @@ PopupMenuAppSwitcherItem.prototype = {
   removeStaleWindowThumbnails: function (windows) {
     for (let i = 0, len = this.appThumbnails.length; i < len; i++) {
       if (this.appThumbnails[i] !== undefined && windows.indexOf(this.appThumbnails[i].metaWindow) === -1) {
-        if (this.appThumbnails[i].thumbnail) {
-          this.box.remove_actor(this.appThumbnails[i].thumbnail.actor);
-          this.appThumbnails[i].thumbnail.destroy();
+        if (this.appThumbnails[i]) {
+          this.box.remove_actor(this.appThumbnails[i].actor);
+          this.appThumbnails[i].destroy();
         }
         _.pullAt(this.appThumbnails, i);
       }
@@ -837,9 +839,12 @@ PopupMenuAppSwitcherItem.prototype = {
   },
 
   destroy: function(){
+    if (this.metaWindowThumbnail && this.metaWindowThumbnail != null) {
+      this.metaWindowThumbnail.destroy();
+    }
     for (let w = 0, len = this.appThumbnails.length; w < len; w++) {
       if (this.appThumbnails[w] !== undefined) {
-        this.appThumbnails[w].thumbnail.destroy(true);
+        this.appThumbnails[w].destroy(true);
       }
     }
   }
@@ -975,36 +980,27 @@ WindowThumbnail.prototype = {
     return false;
   },
 
-  _onFocusChange: function () {
-    if (this._hasFocus()) {
-      this.actor.remove_style_class_name('thumbnail-alerts');
-    }
-  },
-
   _focusWindowChange: function () {
     if (this._hasFocus()) {
-      this.actor.add_style_pseudo_class('selected');
+      this.actor.set_style_pseudo_class('selected');
     } else {
       this.actor.remove_style_pseudo_class('selected');
     }
   },
 
   _hasFocus: function () {
-    if (!this.metaWindow) {
+    if (!this.metaWindow
+      || this.metaWindow.minimized) {
       return false;
     }
 
-    if (this.metaWindow.minimized) {
-      return false;
-    }
-
-    if (this.metaWindow.has_focus()) {
+    if (this.metaWindow.appears_focused) {
       return true;
     }
 
     let transientHasFocus = false;
     this.metaWindow.foreach_transient(function (transient) {
-      if (transient.has_focus()) {
+      if (transient.appears_focused) {
         transientHasFocus = true;
         return false;
       }
@@ -1042,19 +1038,19 @@ WindowThumbnail.prototype = {
   _getThumbnail: function () {
     // Create our own thumbnail if it doesn't exist
     let thumbnail = null;
-    if (this.sizeChangeId && this.muffinWindow) {
-      this.muffinWindow.disconnect(this.sizeChangeId);
+    if (this.muffinWindow) {
+      this.signals.disconnect('size-changed', this.muffinWindow);
     }
     this.muffinWindow = this.metaWindow.get_compositor_private();
     if (this.muffinWindow) {
       let windowTexture = this.muffinWindow.get_texture();
       let [width, height] = windowTexture.get_size();
-      this.sizeChangeId = this.muffinWindow.connect('size-changed', ()=>this._refresh());
+      this.signals.connect(this.muffinWindow, 'size-changed', Lang.bind(this, this._refresh));
       let scale = Math.min(1.0, this.thumbnailWidth / width, this.thumbnailHeight / height);
       thumbnail = new Clutter.Clone({
         source: windowTexture,
         reactive: true,
-        width: width * scale,
+        width: width * scale * this._applet.scaleFactor,
         height: height * scale
       });
     }
@@ -1088,7 +1084,6 @@ WindowThumbnail.prototype = {
     let button = event.get_button();
     if (button === 1 && !this.stopClick && !this.isFavoriteApp) {
       Main.activateWindow(this.metaWindow, global.get_current_time());
-
       this.appSwitcherItem.hoverMenu.close();
     } else if (button === 2 && !this.stopClick) {
       this.handleAfterClick();
@@ -1131,13 +1126,13 @@ WindowThumbnail.prototype = {
         setThumbSize(divider * divideMultiplier, 16);
         return;
       } else {
+        this.thumbnailActor.width = this.thumbnailWidth;
+        this._container.style = 'width: ' + Math.floor(this.thumbnailWidth - 16) + 'px;';
         if (this._applet.verticalThumbs && this._applet.showThumbs) {
           this.thumbnailActor.height = this.thumbnailHeight;
         } else if (this._applet.verticalThumbs) {
           this.thumbnailActor.height = 0;
         }
-        this.thumbnailActor.width = this.thumbnailWidth;
-        this._container.style = 'width: ' + Math.floor(this.thumbnailWidth - 16) + 'px';
 
         this.isFavoriteApp = false;